Azure/Functions/AdDataCollector/requirements.txt
azure-functions
pyodbc
pandas
requests
